move taxonomy field
