+++ src/data/request/query/TaxonomyFilter.js
@@ -2,7 +2,7 @@
 
 import {Record} from 'immutable';
 
-import TaxonomyField from '../../TaxonomyField';
+import TaxonomyField from './TaxonomyField';
 import Taxonomy from '../../Taxonomy';
 
 let defaults = {
