+++ src/data/request/query/builders/TaxonomyFiltersParametersBuilder.js
@@ -3,7 +3,7 @@
 import {Map, List} from 'immutable';
 
 import TaxonomyFilter from '../TaxonomyFilter';
-import TaxonomyField from '../../../TaxonomyField';
+import TaxonomyField from '../TaxonomyField';
 
 export default class TaxonomyFiltersParametersBuilder {
   static build(filters) {
