+++ test/EventsSearchParametersBuilderTest.js
@@ -6,7 +6,7 @@ import {List} from 'immutable';
 
 import Unit from '../src/data/Unit';
 import Taxonomy from '../src/data/Taxonomy';
-import TaxonomyField from '../src/data/TaxonomyField';
+import TaxonomyField from '../src/data/request/query/TaxonomyField';
 import Pagination from '../src/data/request/query/Pagination';
 import VenuesFilter from '../src/data/request/query/VenuesFilter';
 import PerformerFilter from '../src/data/request/query/PerformerFilter';
