+++ test/TaxonomyFiltersParameterBuilderTest.js
@@ -6,7 +6,7 @@ import {List} from 'immutable';
 
 import Taxonomy from '../src/data/Taxonomy';
 import TaxonomyFilter from '../src/data/request/query/TaxonomyFilter';
-import TaxonomyField from '../src/data/TaxonomyField';
+import TaxonomyField from '../src/data/request/query/TaxonomyField';
 import TaxonomyFiltersParametersBuilder from '../src/data/request/query/builders/TaxonomyFiltersParametersBuilder';
 
 describe('Test Taxonomy Filters Parameter Builder', function() {
